added additional tests covering blueprint, runtime, corpus_query, and split the existing test_resources file into a per resource file for future integrations
