+++ tests/resources/__init__.py
@@ -1,0 +1,2 @@
+"""Resource-level SDK tests."""
+
